Add files via upload
In this project, I did an analysis of a Python dataframe, and I wanted to create a menu with options for can see anything that I (or you) want. At the final, I wanted to convert the dataframe from Portuguese-BR to English-EUA, the archives named in portuguese are the original docs. and the archives in english are the new docs.